requirements.txt
#Core dependencies with version constraints to ensure compatibility
numpy==1.26.4
scipy==1.11.4
pandas==2.2.2
tensorflow==2.15.0
ml_dtypes==0.2.0
matplotlib==3.8.4
h5py==3.10.0
scikit-learn==1.3.2
ta==0.11.0
tensorboard==2.15.2
contourpy==1.2.1

fastapi-mail==1.5.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.0.0  # Using newer version compatible with updated supabase stack
pydantic[email]  # For EmailStr support
sqlalchemy==2.0.23
python-multipart==0.0.6
fastapi-cli

# Brevo email service
sib-api-v3-sdk==7.6.0

# Environment variables
python-dotenv==1.0.0

# Supabase dependencies - using newer compatible versions
supabase==2.16.0
postgrest==1.1.1
storage3==0.12.0
gotrue==2.12.3
realtime==2.5.3
httpx==0.28.1

# Database
psycopg==3.2.9
psycopg-binary==3.2.9
alembic==1.13.1
aiosqlite==0.21.0

# Utilities
requests==2.31.0
typing-extensions>=4.6.0
email-validator  # Required for pydantic email validation

# Testing dependencies
pytest==8.4.1
pytest-cov==4.1.0
pytest-asyncio==1.0.0

memory-profiler==0.61.0
celery==5.5.3
python-dateutil==2.9.0
pydantic-settings==2.8.1
redis==6.2.0
motor==3.7.1
pymongo==4.13.2
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
APScheduler==3.11.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cycler==0.12.1
dateparser==1.2.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi-cli==0.0.7
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
httptools==0.6.4
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
namex==0.1.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-binance==1.0.29
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
rich==14.0.0
rich-toolkit==0.14.8
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
xgboost==3.0.2
yarl==1.20.1
